+++ src/main/java/com/project/zighang/domain/company/controller/CompanyController.java
@@ -6,7 +6,7 @@ import com.project.zighang.domain.company.enumerate.CompanyType;
 import com.project.zighang.domain.company.enumerate.JobGroup;
 import com.project.zighang.domain.company.service.CompanyQueryService;
 import com.project.zighang.global.exception.Success;
-import com.project.zighang.global.template.RspTemplate;
+import com.project.zighang.global.exception.template.RspTemplate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.web.bind.annotation.*;
