+++ src/main/java/com/project/zighang/global/client/clova/dto/ClovaOcrResponse.java
@@ -1,4 +1,4 @@
-package com.project.zighang.global.dto;
+package com.project.zighang.global.client.clova.dto;
 
 import java.util.List;
