+++ src/main/java/com/project/zighang/global/client/email/service/SmtpEmailSender.java
@@ -1,5 +1,6 @@
-package com.project.zighang.global.service;
+package com.project.zighang.global.client.email.service;
 
+import com.project.zighang.global.client.email.EmailSender;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class SmtpEmailSender implements EmailSender{
+public class SmtpEmailSender implements EmailSender {
 
     private final JavaMailSender javaMailSender;
