+++ src/main/java/com/project/zighang/global/client/objectStorage/PresignedUrlReader.java
@@ -1,0 +1,7 @@
+package com.project.zighang.global.client.objectStorage;
+
+import com.project.zighang.global.client.objectStorage.dto.PreSignedUrlResponse;
+
+public interface PresignedUrlReader {
+    PreSignedUrlResponse getPreSignedUrl(String prefix, String originalFileName);
+}
