--- src/main/java/com/project/zighang/global/service/PresignedUrlReader.java
@@ -1,7 +1,0 @@
-package com.project.zighang.global.service;
-
-import com.project.zighang.global.dto.PreSignedUrlResponse;
-
-public interface PresignedUrlReader {
-    PreSignedUrlResponse getPreSignedUrl(String prefix, String originalFileName);
-}
